Fix NPC exp share overflow
git-svn-id: svn://svn.eoserv.net/eoserv/trunk@303 da1c5d5e-72c9-4536-8af9-af1c9134b0e6

+++ src/npc.hpp
@@ -25,7 +25,7 @@
 struct NPC_Opponent
 {
 	Character *attacker;
-	unsigned short damage;
+	int damage;
 	double last_hit;
 };
